feat: add function to handler deque list

+++ ft_deque_handler.h
@@ -6,25 +6,24 @@
 /*   By: jtertuli <jtertuli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/18 09:09:26 by jtertuli          #+#    #+#             */
-/*   Updated: 2025/08/19 10:29:32 by jtertuli         ###   ########.fr       */
+/*   Updated: 2025/08/20 12:14:38 by jtertuli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 
 #ifndef FT_DEQUE_HANDLER_H
 # define FT_DEQUE_HANDLER_H
 
 # include "ft_node_handler.h"
 
-typedef struct s_deque t_deque;
+typedef struct s_deque	t_deque;
 
 typedef struct s_deque {
 	t_node	*top;
 	t_node	*bottom;
-	int		size;
+	size_t	size;
 }	t_deque;
 
-t_deque *ft_deque_create(void);
+t_deque	*ft_deque_create(void);
 void	ft_deque_push_top(t_deque *deque, int value);
 void	ft_deque_push_bottom(t_deque *deque, int value);
 void	ft_deque_pop_top(t_deque *deque);
